+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Body-frame acceleration in `AircraftState`
 - Working sideslip angle indicator in on ADI
 - Argument descriptions for generated code
+- Pack maker for [MBDyn](https://www.mbdyn.org/)
 
 ### Changed